+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Ability to use timeOffset only when local system is using non-UTC timezone
 - Support to replace existing logging libraries with ease
 
+### Fixed
+- Attach stream only when logging client connected
+
 ## [1.0.0-beta.2]
 ### Added
 - Added support to manually set bulk size and enable/disable bulk dispatch